Start work on ffixed_t template, to cover fixed_t, fixed64_t and shint64_t.

+++ src/doom/p_floor.cpp
@@ -395,7 +395,7 @@ EV_DoFloor
 	    if (floor->floordestheight > sec->ceilingheight)
 		floor->floordestheight = sec->ceilingheight;
 	    floor->floordestheight -= (8*FRACUNIT)*
-		(floortype == raiseFloorCrush);
+               (int)(floortype == raiseFloorCrush);
 	    break;
 
 	  case raiseFloorTurbo:
